Move footer components to their own folder
