+++ next/components/footer/footer.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
 
-import { SocialShare } from "@/components/social-share";
+import { SocialShare } from "@/components/footer/social-share";
 import type { Menu, MenuItem, MenuItemOptions } from "@/lib/zod/menu";
 import Facebook from "@/styles/icons/facebook.svg";
 import LinkedIn from "@/styles/icons/linkedin.svg";
