+++ next/components/layout.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from "next-i18next";
 import React from "react";
 import clsx from "clsx";
 
-import { Footer } from "@/components/footer";
+import { Footer } from "@/components/footer/footer";
 import { Header } from "@/components/header/header";
 import {
   PreviewBanner,
